Rewrote the division operations for the benefit of SMT-RAT.
Rewrote both the 2 stage problems, flat and layered. Have a draft
version of a function that will automatically rewrite an SMT file like
this to replace variables bound to quotients with equations that use the
same variables in multiplication with the divisor.

SMT-RAT will get a floating point exception for divisions unless the
quotient is a constant. They claim that this is reasonable because they
can't prove that the divisor isn't zero.