fix issues related to private fields and floats
pydantic was not assigning values that we were expecting:
EncodedModel(_formula=formula) was leaving _formula at default
instead of at whatever the value of formula was. This commit
holds a fix for the specific case of EncodedModel but a more
exhaustive review of the private fields should be considered.

Additionally this commit converts the constants for negative
and positive infinity to be proper floats instead of strings.
It also tells pydantic to allow infinity values for floats
(though that should be the default behavior). This should fix
some bugs related to boxes with bounds at infinity (failing to
serialize etc)